+++ ecd-20200930/src/Models/DescribeDesktopGroupSessionsResponseBody.php
@@ -1,0 +1,110 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Ecd\V20200930\Models;
+
+use AlibabaCloud\SDK\Ecd\V20200930\Models\DescribeDesktopGroupSessionsResponseBody\sessions;
+use AlibabaCloud\Tea\Model;
+
+class DescribeDesktopGroupSessionsResponseBody extends Model
+{
+    /**
+     * @description A pagination token. It can be used in the next request to retrieve a new page of results.
+     *
+     * @example caeba0bbb2
+     *
+     * @var string
+     */
+    public $nextToken;
+
+    /**
+     * @description The request ID.
+     *
+     * @example D0920845-7359-59FC-9899-B****
+     *
+     * @var string
+     */
+    public $requestId;
+
+    /**
+     * @description The sessions.
+     *
+     * @var sessions[]
+     */
+    public $sessions;
+
+    /**
+     * @description The total number of sessions.
+     *
+     * @example 15
+     *
+     * @var int
+     */
+    public $totalCount;
+    protected $_name = [
+        'nextToken'  => 'NextToken',
+        'requestId'  => 'RequestId',
+        'sessions'   => 'Sessions',
+        'totalCount' => 'TotalCount',
+    ];
+
+    public function validate()
+    {
+    }
+
+    public function toMap()
+    {
+        $res = [];
+        if (null !== $this->nextToken) {
+            $res['NextToken'] = $this->nextToken;
+        }
+        if (null !== $this->requestId) {
+            $res['RequestId'] = $this->requestId;
+        }
+        if (null !== $this->sessions) {
+            $res['Sessions'] = [];
+            if (null !== $this->sessions && \is_array($this->sessions)) {
+                $n = 0;
+                foreach ($this->sessions as $item) {
+                    $res['Sessions'][$n++] = null !== $item ? $item->toMap() : $item;
+                }
+            }
+        }
+        if (null !== $this->totalCount) {
+            $res['TotalCount'] = $this->totalCount;
+        }
+
+        return $res;
+    }
+
+    /**
+     * @param array $map
+     *
+     * @return DescribeDesktopGroupSessionsResponseBody
+     */
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['NextToken'])) {
+            $model->nextToken = $map['NextToken'];
+        }
+        if (isset($map['RequestId'])) {
+            $model->requestId = $map['RequestId'];
+        }
+        if (isset($map['Sessions'])) {
+            if (!empty($map['Sessions'])) {
+                $model->sessions = [];
+                $n               = 0;
+                foreach ($map['Sessions'] as $item) {
+                    $model->sessions[$n++] = null !== $item ? sessions::fromMap($item) : $item;
+                }
+            }
+        }
+        if (isset($map['TotalCount'])) {
+            $model->totalCount = $map['TotalCount'];
+        }
+
+        return $model;
+    }
+}